requirements.txt
asgiref==3.3.1
automium==0.2.6
automium-web==0.1.1
click==7.1.2
common==0.1.2
Django==3.1.4
django-cors-headers==3.6.0
django-extensions==3.1.1
django-filter==2.4.0
djangorestframework==3.12.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
langdetect==1.0.8
legacy==0.1.6
MarkupSafe==1.1.1
Pillow==8.1.0
psycopg2==2.8.6
pytz==2020.4
reportlab==3.5.59
six==1.15.0
sqlparse==0.4.1
Werkzeug==1.0.1
wfastcgi==3.0.0
